refactor(ff-encode): move StreamCopyTrimmer impl from trim/mod.rs to trim/trimmer.rs
